backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
python-multipart
pydantic
requests
dotenv